Set a higher priority for the hook
I was having problems with the leaflet map not being initialized because the javascript was not being included on the page because the hook priority of this module was 100.1 and the priority of the TemplateEngineFactory::hookRender() was 100.3.

I think 100 is the default priority.  I think we can safely raise the priority of this module to 200 so that it runs after TemplateEngineFactory and others.

For more details on the problem I having see https://processwire.com/talk/topic/9745-module-leaflet-map/?do=findComment&comment=140014

Hope that helps!